Boat Navigation Update, Enemies and data,  State Controllers Galore
- New scriptable objects for Enemies that implement data based on their move set
- Game Section Manager now has references to the new enemy types for spawning.
- Introduced an additional debug option that shows debug if a Section Object was selected. This has been used for showing where Crocodile enemies will appear and jump on the boat.
- Removed a few Injection dependencies in replacement for a static reference. This is temporary.
- Introduced MonoTimeBehaviour, a class based on MonoBehaviour for objects affected by the game being paused.
- Replaced/renamed the Boat_Character class to CharacterStateController. Done the same with the Player_Controller, which is now PlayerStateController.
- Added animations for enemies
- Updated Boat Movement to allow movement to the sides of the boat. Additionally, characters have checks for whether or not they can enter spaces on the boat, sides of the boat or both.

+++ Assets/Scripts/River Objects/River_Obstacle.cs
@@ -53,5 +53,5 @@ public class River_Obstacle : River_Object
 [Serializable]
 public class ObstacleData
 {
-    public float ImpactDamage = 1f;
+    public int ImpactDamage = 1;
 }
